start transfer xsc audit integration test to git audit
