+++ audit_test.go
@@ -817,7 +817,11 @@ type auditCommandTestParams struct {
 
 // run audit command with different flags and params for integration tests
 func testAuditCommand(t *testing.T, testCli *coreTests.JfrogCli, params auditCommandTestParams) (string, error) {
-	args := []string{"audit"}
+	args := append([]string{"audit"}, getAuditCmdArgs(params)...)
+	return testCli.RunCliCmdWithOutputs(t, args...)
+}
+
+func getAuditCmdArgs(params auditCommandTestParams) (args []string) {
 	if len(params.WorkingDirsToScan) > 0 {
 		args = append(args, "--working-dirs="+strings.Join(params.WorkingDirsToScan, ","))
 	}
@@ -843,5 +847,5 @@ func testAuditCommand(t *testing.T, testCli *coreTests.JfrogCli, params auditCom
 	if params.WithVuln {
 		args = append(args, "--vuln")
 	}
-	return testCli.RunCliCmdWithOutputs(t, args...)
+	return args
 }
